+++ examples/basics/factories.ts
@@ -1,4 +1,4 @@
-import { ActionsFactory } from "../src";
+import { ActionsFactory } from "../../src";
 import { config } from "./config";
 
 /** @desc this factory is for producing actions - handlers of the incoming socket.io events */