+++ examples/basics/generate-client.ts
@@ -1,10 +1,10 @@
 import { writeFile } from "node:fs/promises";
-import { Integration } from "../src";
+import { Integration } from "../../src";
 import { actions } from "./actions";
 import { config } from "./config";
 
 await writeFile(
-  "example/example-client.ts",
+  "examples/basics/example-client.ts",
   new Integration({ config, actions }).print(),
   "utf-8",
 );