+++ examples/basics/index.ts
@@ -1,5 +1,5 @@
 import http from "node:http";
-import { attachSockets } from "../src";
+import { attachSockets } from "../../src";
 import { actions } from "./actions";
 import { config } from "./config";
 import { Server } from "socket.io";